unit test added for all distances

+++ src/utils.rs
@@ -18,9 +18,9 @@ pub fn derivate(x: &Vec<Vec<f64>>) -> Vec<Vec<f64>> {
 const WEIGHT_MAX: f64 = 1.0;
 pub fn dtw_weights(len: usize, g: f64) -> Vec<f64> {
     let mut weights = vec![0.0; len];
-    let half_len = len as f64 / 2.0;
+    let half_len = len / 2;
     for i in 0..len {
-        weights[i] = WEIGHT_MAX / (1.0 + std::f64::consts::E.powf(-g * (i as f64 - half_len)));
+        weights[i] = WEIGHT_MAX / (1.0 + std::f64::consts::E.powf(-g * (i as f64 - half_len as f64)));
     }
     weights
 }
